I think I did it

+++ lib/src/generalTypes/age.dart
@@ -11,7 +11,7 @@ part 'age.g.dart';
 class Age {
   String id;
   List<Extension> extension;
-  Decimal value;
+  double value;
   AmountComparator comparator;
   String unit;
   FhirUri system;
