+++ lib/src/generalTypes/distance.dart
@@ -11,7 +11,7 @@ part 'distance.g.dart';
 class Distance {
   String id;
   List<Extension> extension;
-  Decimal value;
+  double value;
   AmountComparator comparator;
   String unit;
   FhirUri system;
